requirements.txt
albumentations==0.4.3
asn1crypto==0.24.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
bokeh==1.4.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
colorama==0.4.1
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
efficientnet-pytorch==0.5.1
entrypoints==0.3
et-xmlfile==1.0.1
idna==2.8
imageio==2.5.0
imgaug==0.2.6
imutils==0.5.3
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.5.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
munch==2.5.0
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
notebook==5.7.8
numpy==1.16.4
olefile==0.46
opencv-python==4.1.2.30
opencv-python-headless==4.1.1.26
openpyxl==3.0.0
packaging==19.2
pandas==0.24.2
pandocfilters==1.4.2
parso==0.5.0
pickleshare==0.7.5
Pillow==6.0.0
plotly==3.10.0
pretrainedmodels==0.7.4
prometheus-client==0.7.1
prompt-toolkit==2.0.9
pycparser==2.19
Pygments==2.4.2
pyOpenSSL==19.0.0
pyparsing==2.4.0
pyrsistent==0.14.11
PySocks==1.7.0
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
pywinpty==0.5.5
PyYAML==5.1.2
pyzmq==18.0.0
qtconsole==4.5.1
requests==2.22.0
retrying==1.3.3
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.10.0
segmentation-models-pytorch==0.1.0
Send2Trash==1.5.0
Shapely==1.6.4.post1
six==1.12.0
terminado==0.8.2
testpath==0.4.2
torch==1.1.0
torchvision==0.3.0
tornado==6.0.3
tqdm==4.32.2
traitlets==4.3.2
urllib3==1.24.2
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==1.2.0
